Prevent the dictionaries from updating G__struct until it is actually initialized.  (In particular for dictionary in the libraries hard linked to the exectable)
git-svn-id: http://root.cern.ch/svn/root/trunk@27175 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/cint7/src/disp.cxx
@@ -2561,8 +2561,8 @@ extern "C" int G__display_class(FILE *fout, char *name, int base, int start)
    }
    if (G__more(fout, msg)) return(1);
    sprintf(msg
-           , " (tagnum=%d,voffset=%d,isabstract=%d,parent=%d,gcomp=%d:%d,d21=~cd=%x)"
-           , tagnum , G__struct.virtual_offset[tagnum]
+           , " (tagnum=%d,voffset=%ld,isabstract=%d,parent=%d,gcomp=%d:%d,d21=~cd=%x)"
+           , tagnum , (long)G__struct.virtual_offset[tagnum]
            , G__struct.isabstract[tagnum] , G__struct.parent_tagnum[tagnum]
            , G__struct.globalcomp[tagnum], G__struct.iscpplink[tagnum]
            , G__struct.funcs[tagnum]);
